refactor shared code into separate files
