+++ external/fv3fit/fv3fit/keras/__init__.py
@@ -1,3 +1,11 @@
-from ._models.models import History
-from ._models.recurrent import PureKerasModel, StepwiseModel
 from . import _models
+from ._models.shared.dense_network import DenseNetwork, DenseNetworkConfig
+from ._models.shared.pure_keras import PureKerasModel
+from ._models.shared.training_loop import (
+    TrainingLoopConfig,
+    EpochResult,
+    EpochLossHistory,
+    History,
+)
+from ._models.shared.utils import get_input_vector
+from ._models.recurrent import StepwiseModel
