+++ external/fv3fit/fv3fit/keras/_models/models.py
@@ -1,7 +1,5 @@
 from typing import (
-    Callable,
     Sequence,
-    Tuple,
     Iterable,
     Mapping,
     Union,
@@ -36,11 +34,11 @@ from ..._shared.config import (
 import numpy as np
 import os
 from ._filesystem import get_dir, put_dir
-from ._sequences import _XyArraySequence, _ThreadedSequencePreLoader
+from ._sequences import _XyArraySequence
 from .normalizer import LayerStandardScaler
 from .loss import get_weighted_mse, get_weighted_mae
-from .shared import DenseNetworkConfig
-from loaders.batches import Take, shuffle
+from .shared import DenseNetworkConfig, TrainingLoopConfig, EpochResult
+from loaders.batches import Take
 import yaml
 from vcm import safe
 
@@ -49,77 +47,6 @@ logger = logging.getLogger(__file__)
 
 MODEL_DIRECTORY = "model_data"
 KERAS_CHECKPOINT_PATH = "model_checkpoints"
-
-# Description of the training loss progression over epochs
-# Outer array indexes epoch, inner array indexes batch (if applicable)
-EpochLossHistory = Sequence[Sequence[Union[float, int]]]
-History = Mapping[str, EpochLossHistory]
-
-
-@dataclasses.dataclass
-class EpochResult:
-    """
-    Attributes:
-        epoch: count of epoch (starts at zero)
-        history: return value of `model.fit` from each batch
-    """
-
-    epoch: int
-    history: Sequence[tf.keras.callbacks.History]
-
-
-@dataclasses.dataclass
-class TrainingLoopConfig:
-    """
-    epochs: number of times to run through the batches when training
-    workers: number of workers for parallelized loading of batches fed into
-        training, if 1 uses serial loading instead
-    max_queue_size: max number of batches to hold in the parallel loading queue
-    keras_batch_size: actual batch_size to pass to keras model.fit,
-        independent of number of samples in each data batch in batches
-    """
-
-    epochs: int = 3
-    workers: int = 1
-    max_queue_size: int = 8
-    keras_batch_size: int = 16
-
-    def fit_loop(
-        self,
-        model: tf.keras.Model,
-        Xy: Sequence[Tuple[np.ndarray, np.ndarray]],
-        validation_data: Optional[Tuple[np.ndarray, np.ndarray]] = None,
-        callbacks: Iterable[Callable[[EpochResult], None]] = (),
-    ) -> None:
-        """
-        Args:
-            model: keras model to train
-            Xy: sequence of data batches to be passed to `model.fit`
-            validation_data: passed as `validation_data` argument to `model.fit`
-            callbacks: if given, these will be called at the end of each epoch
-        """
-        for i_epoch in range(self.epochs):
-            Xy = shuffle(Xy)
-            if self.workers > 1:
-                Xy = _ThreadedSequencePreLoader(
-                    Xy, num_workers=self.workers, max_queue_size=self.max_queue_size
-                )
-            history = []
-            for i_batch, (X, y) in enumerate(Xy):
-                logger.info(
-                    f"Fitting on batch {i_batch + 1} of {len(Xy)}, "
-                    f"of epoch {i_epoch}..."
-                )
-                history.append(
-                    model.fit(
-                        X,
-                        y,
-                        validation_data=validation_data,
-                        batch_size=self.keras_batch_size,
-                    )
-                )
-            for callback in callbacks:
-                callback(EpochResult(epoch=i_epoch, history=tuple(history)))
 
 
 @dataclasses.dataclass
