+++ external/fv3fit/fv3fit/keras/_models/shared/pure_keras.py
@@ -1,4 +1,3 @@
-from fv3fit._shared.packer import ArrayPacker
 from fv3fit._shared import (
     Predictor,
     io,
@@ -6,118 +5,12 @@ from fv3fit._shared import (
     SAMPLE_DIM_NAME,
     stack_non_vertical,
 )
-from fv3fit._shared.config import RegularizerConfig
 import tensorflow as tf
-from typing import Any, Dict, Hashable, Iterable, Optional, Sequence
+from typing import Any, Dict, Hashable, Iterable
 import xarray as xr
 import os
-from ._filesystem import get_dir, put_dir
+from .._filesystem import get_dir, put_dir
 import yaml
-import dataclasses
-import tensorflow_addons as tfa
-
-
-def get_input_vector(
-    packer: ArrayPacker, n_window: Optional[int] = None, series: bool = True,
-):
-    """
-    Given a packer, return a list of input layers with one layer
-    for each input used by the packer, and a list of output tensors which are
-    the result of packing those input layers.
-
-    Args:
-        packer
-        n_window: required if series is True, number of timesteps in a sample
-        series: if True, returned inputs have shape [n_window, n_features], otherwise
-            they are 1D [n_features] arrays
-    """
-    features = [packer.feature_counts[name] for name in packer.pack_names]
-    if series:
-        if n_window is None:
-            raise TypeError("n_window is required if series is True")
-        input_layers = [
-            tf.keras.layers.Input(shape=[n_window, n_features])
-            for n_features in features
-        ]
-    else:
-        input_layers = [
-            tf.keras.layers.Input(shape=[n_features]) for n_features in features
-        ]
-    packed = tf.keras.layers.Concatenate()(input_layers)
-    return input_layers, packed
-
-
-@dataclasses.dataclass
-class DenseNetwork:
-    """
-    Attributes:
-        output: final output of dense network
-        hidden_outputs: consecutive outputs of hidden layers in dense network
-    """
-
-    output: tf.Tensor
-    hidden_outputs: Sequence[tf.Tensor]
-
-
-@dataclasses.dataclass
-class DenseNetworkConfig:
-    width: int = 8
-    depth: int = 3
-    kernel_regularizer: RegularizerConfig = dataclasses.field(
-        default_factory=lambda: RegularizerConfig("none")
-    )
-    gaussian_noise: float = 0.0
-    spectral_normalization: bool = False
-    output_activity_regularizer: RegularizerConfig = dataclasses.field(
-        default_factory=lambda: RegularizerConfig("none")
-    )
-
-    def build(
-        self, x_in: tf.Tensor, n_features_out: int, label: str = ""
-    ) -> DenseNetwork:
-        """
-        Take an input tensor to a dense network and return the result of a dense
-        network's prediction, as a tensor.
-
-        Can be used within code that builds a larger neural network. This should
-        take in and return normalized values.
-
-        Args:
-            x_in: input tensor whose last dimension is the feature dimension
-            n_features_out: dimensionality of last (feature) dimension of output
-            config: configuration for dense network
-            label: inserted into layer names, if this function is used multiple times
-                to build one network you must provide a different label each time
-        
-        Returns:
-            tensor resulting from the requested dense network
-        """
-        hidden_outputs = []
-        x = x_in
-        for i in range(self.depth - 1):
-            if self.gaussian_noise > 0.0:
-                x = tf.keras.layers.GaussianNoise(
-                    self.gaussian_noise, name=f"gaussian_noise_{label}_{i}"
-                )(x)
-            hidden_layer = tf.keras.layers.Dense(
-                self.width,
-                activation=tf.keras.activations.relu,
-                kernel_regularizer=self.kernel_regularizer.instance,
-                name=f"hidden_{label}_{i}",
-            )
-            if self.spectral_normalization:
-                hidden_layer = tfa.layers.SpectralNormalization(
-                    hidden_layer, name=f"spectral_norm_{label}_{i}"
-                )
-            x = hidden_layer(x)
-            hidden_outputs.append(x)
-        output = tf.keras.layers.Dense(
-            n_features_out,
-            activation="linear",
-            # activity_regularizer=self.output_activity_regularizer.instance,
-            name=f"dense_network_{label}_output",
-        )(x)
-        return DenseNetwork(hidden_outputs=hidden_outputs, output=output)
 
 
 @io.register("all-keras")
